# openapi.json
{"openapi":"3.1.0","info":{"title":"Aura Research - Document Analysis System (MinIO + Celery)","version":"0.1.0"},"paths":{"/health":{"get":{"tags":["system"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/upload":{"post":{"tags":["documents"],"summary":"Upload File","operationId":"upload_file_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/status/{task_id}":{"get":{"tags":["documents"],"summary":"Status","operationId":"status_status__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/documents/{doc_id}":{"get":{"tags":["documents"],"summary":"Get Document Details","description":"Recupera los detalles de un documento procesado (categoría, entidades, chunks).","operationId":"get_document_details_documents__doc_id__get","parameters":[{"name":"doc_id","in":"path","required":true,"schema":{"type":"string","title":"Doc Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["documents"],"summary":"Delete Document","description":"Elimina un documento y sus vectores asociados.","operationId":"delete_document_documents__doc_id__delete","parameters":[{"name":"doc_id","in":"path","required":true,"schema":{"type":"string","title":"Doc Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search":{"get":{"tags":["search"],"summary":"Search","description":"Búsqueda sin cambios (usa Qdrant).","operationId":"search_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","minLength":3,"title":"Q"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":20,"minimum":1,"default":5,"title":"Limit"}},{"name":"entity_filter","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Entity Filter"}},{"name":"category_filter","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category Filter"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat":{"post":{"tags":["search"],"summary":"Chat","description":"Endpoint RAG: Responde preguntas usando el agente LangGraph (con Tools).","operationId":"chat_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tools/":{"get":{"tags":["tools"],"summary":"List Tools","operationId":"list_tools_tools__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ToolConfigRead"},"type":"array","title":"Response List Tools Tools  Get"}}}}}},"post":{"tags":["tools"],"summary":"Create Tool","operationId":"create_tool_tools__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolConfigCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolConfigRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tools/{tool_id}":{"get":{"tags":["tools"],"summary":"Get Tool","operationId":"get_tool_tools__tool_id__get","parameters":[{"name":"tool_id","in":"path","required":true,"schema":{"type":"integer","title":"Tool Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolConfigRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["tools"],"summary":"Update Tool","operationId":"update_tool_tools__tool_id__put","parameters":[{"name":"tool_id","in":"path","required":true,"schema":{"type":"integer","title":"Tool Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolConfigUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolConfigRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["tools"],"summary":"Delete Tool","operationId":"delete_tool_tools__tool_id__delete","parameters":[{"name":"tool_id","in":"path","required":true,"schema":{"type":"integer","title":"Tool Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_file_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_file_upload_post"},"ChatRequest":{"properties":{"query":{"type":"string","title":"Query"}},"type":"object","required":["query"],"title":"ChatRequest"},"ChatResponse":{"properties":{"answer":{"type":"string","title":"Answer"},"sources":{"items":{"type":"object"},"type":"array","title":"Sources"}},"type":"object","required":["answer","sources"],"title":"ChatResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SearchResponse":{"properties":{"query":{"type":"string","title":"Query"},"results":{"items":{"$ref":"#/components/schemas/SearchResult"},"type":"array","title":"Results"}},"type":"object","required":["query","results"],"title":"SearchResponse"},"SearchResult":{"properties":{"text":{"type":"string","title":"Text"},"score":{"type":"number","title":"Score"},"metadata":{"type":"object","title":"Metadata"}},"type":"object","required":["text","score","metadata"],"title":"SearchResult"},"ToolConfigCreate":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"api_url":{"type":"string","title":"Api Url"},"method":{"type":"string","title":"Method","default":"GET"},"headers":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Headers","default":{}},"auth_config":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Auth Config","default":{}},"system_prompt":{"type":"string","title":"System Prompt"},"parameter_schema":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Parameter Schema","default":{}}},"type":"object","required":["name","description","api_url","system_prompt"],"title":"ToolConfigCreate"},"ToolConfigRead":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"api_url":{"type":"string","title":"Api Url"},"method":{"type":"string","title":"Method","default":"GET"},"headers":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Headers","default":{}},"auth_config":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Auth Config","default":{}},"system_prompt":{"type":"string","title":"System Prompt"},"parameter_schema":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Parameter Schema","default":{}},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","description","api_url","system_prompt","id"],"title":"ToolConfigRead"},"ToolConfigUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"api_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Api Url"},"method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Method"},"headers":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Headers"},"auth_config":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Auth Config"},"system_prompt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"System Prompt"},"parameter_schema":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Parameter Schema"}},"type":"object","title":"ToolConfigUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}